+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix duplicate player head rendering on pvp gui on remote
 - Disable "Kick from party" button when you're not a party owner
 - Fix pvp locking on party of 2 players where 1 is defeated and the other one leaves the game
+- Fix double party name change input dialog on remote
 
 ## [0.8.2] 2026-07-26